refactor: 한 객체에 모여있는 각기 다른 기능들을 인터페이스를 사용하여 분리
