+++ src/main/java/webserver/MainHandler.java
@@ -2,7 +2,7 @@ package webserver;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import webserver.handler.RequestHandler;
+import webserver.handler.HandlerMapper;
 import http.requestMessage.Request;
 import http.responseMessage.Response;
 
@@ -26,7 +26,7 @@ public class MainHandler implements Runnable {
             final String requestMessage = getRequestMessage(in);
 
             final Request request = new Request(requestMessage);
-            final RequestHandler handler = new RequestHandler(request);
+            final HandlerMapper handler = new HandlerMapper(request);
             final Response response = handler.handle();
 
             final BufferedOutputStream bos = new BufferedOutputStream(out);
