+++ src/main/java/webserver/handler/Handler.java
@@ -1,4 +1,8 @@
 package webserver.handler;
 
+import http.responseMessage.Response;
+
 public interface Handler {
+
+    Response handle();
 }
